ongoing/LegoReader/Python/requirements.txt
astroid==2.2.5
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.3
opencv-python==4.1.0.25
paho-mqtt==1.4.0
pylint==2.3.1
six==1.12.0
typed-ast==1.3.4
wrapt==1.11.1
